+++ src/calendarjs.js
@@ -270,7 +270,6 @@
  * @property    {boolean}   showDayNumberOrdinals                       States if the day ordinal values should be shown (defaults to true).  
  * @property    {boolean}   dragAndDropForEventsEnabled                 States if dragging and dropping events around the days of the month is enabled (defaults to true).
  * @property    {number}    maximumEventsPerDayDisplay                  The maximum number of events that should be display per day in the main calendar display (defaults to 3, 0 disables it).
- * @property    {number}    extraSelectableYearsAhead                   The number of extra years ahead that are selectable in the drop down (defaults to 100).
  * @property    {boolean}   exportEventsEnabled                         States if exporting events is enabled (defaults to true).
  * @property    {boolean}   manualEditingEnabled                        States if adding, editing, dragging and removing events is enabled (defaults to true).
  * @property    {boolean}   showTimesInMainCalendarEvents               States if the time should be shown on the main calendar view events (defaults to false).
@@ -318,6 +317,8 @@
  * @property    {boolean}   applyCssToEventsNotInCurrentMonth           States if extra CSS should be applied to events that are not in the current (on the main display, defaults to true).
  * @property    {boolean}   addYearButtonsInDatePickerMode              States if the year-jumping buttons should be added in DatePicker mode (defaults to false).
  * @property    {number[]}  workingDays                                 States the day numbers that that are considered working days (defaults to [ 0, 1, 2, 3, 4, 5, 6 ], Mon=0, Sun=6).
+ * @property    {number}    minimumYear                                 The minimum year that can be shown in the Calendar (defaults to 1900).
+ * @property    {number}    maximumYear                                 The maximum year that can be shown in the Calendar (defaults to 2100).
  */
 
 
@@ -470,8 +471,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _copiedEventDetails = null,
         _copiedEventDetails_Cut = false,
         _previousDaysVisibleBeforeSingleDayView = [],
-        _year_Minimum = 1900,
-        _year_Maximum = null,
         _elementID_Day = "day-",
         _elementID_Month = "month-",
         _elementID_WeekDay = "week-day-",
@@ -1645,9 +1644,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      */
 
     function buildYearSelectorDropDown( container ) {
-        var date = new Date( _year_Minimum, 1, 1 ),
-            dateCurrent = new Date(),
-            dateYearsTotal = ( dateCurrent.getFullYear() - date.getFullYear() ) + _options.extraSelectableYearsAhead;
+        var yearDate = new Date( _options.minimumYear, 1, 1 );
 
         _element_HeaderDateDisplay_YearSelector = createElement( "div", "years-drop-down" );
         container.appendChild( _element_HeaderDateDisplay_YearSelector );
@@ -1655,12 +1652,15 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_HeaderDateDisplay_YearSelector_Contents = createElement( "div", "contents custom-scroll-bars" );
         _element_HeaderDateDisplay_YearSelector.appendChild( _element_HeaderDateDisplay_YearSelector_Contents );
 
-        for ( var yearIndex = 0; yearIndex < dateYearsTotal; yearIndex++ ) {
-            buildYearSelectorDropDownYear( date.getFullYear() );
-            moveDateForwardYear( date );
+        while ( true ) {
+            buildYearSelectorDropDownYear( yearDate.getFullYear() );
+            moveDateForwardYear( yearDate );
+
+            if ( yearDate.getFullYear() > _options.maximumYear ) {
+                break;
+            }
         }
 
-        _year_Maximum = date.getFullYear() - 1;
         _element_HeaderDateDisplay_Text.onclick = showYearSelectorDropDownMenu;
     }
 
@@ -8450,7 +8450,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             var newDate = new Date( date );
 
             if ( !doDatesMatch( _currentDate, newDate ) ) {
-                if ( newDate.getFullYear() >= _year_Minimum && newDate.getFullYear() <= _year_Maximum ) {
+                if ( newDate.getFullYear() >= _options.minimumYear && newDate.getFullYear() <= _options.maximumYear ) {
                     build( newDate );
                     triggerOptionsEventWithData( "onSetDate", newDate );
                 }
@@ -8487,7 +8487,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 newDateAllowed = isDateValidForDatePicker( newDate );
             
             if ( newDateAllowed && !doDatesMatch( newDate, _currentDateForDatePicker ) ) {
-                if ( newDate.getFullYear() >= _year_Minimum && newDate.getFullYear() <= _year_Maximum ) {
+                if ( newDate.getFullYear() >= _options.minimumYear && newDate.getFullYear() <= _options.maximumYear ) {
                     newDate.setHours( 0, 0, 0, 0 );
 
                     hideDatePickerMode();
@@ -8541,7 +8541,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             var previousMonth = new Date( _currentDate );
             previousMonth.setMonth( previousMonth.getMonth() - 1 );
 
-            if ( previousMonth.getFullYear() >= _year_Minimum ) {
+            if ( previousMonth.getFullYear() >= _options.minimumYear ) {
                 build( previousMonth );
                 triggerOptionsEventWithData( "onPreviousMonth", previousMonth );
             }
@@ -8557,7 +8557,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             var nextMonth = new Date( _currentDate );
             nextMonth.setMonth( nextMonth.getMonth() + 1 );
 
-            if ( nextMonth.getFullYear() <= _year_Maximum ) {
+            if ( nextMonth.getFullYear() <= _options.maximumYear ) {
                 build( nextMonth );
                 triggerOptionsEventWithData( "onNextMonth", nextMonth );
             }
@@ -8569,7 +8569,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             var previousYear = new Date( _currentDate );
             previousYear.setFullYear( previousYear.getFullYear() - 1 );
     
-            if ( previousYear.getFullYear() >= _year_Minimum ) {
+            if ( previousYear.getFullYear() >= _options.minimumYear ) {
                 build( previousYear );
                 triggerOptionsEventWithData( "onPreviousYear", previousYear );
             }
@@ -8581,7 +8581,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             var nextYear = new Date( _currentDate );
             nextYear.setFullYear( nextYear.getFullYear() + 1 );
 
-            if ( nextYear.getFullYear() <= _year_Maximum ) {
+            if ( nextYear.getFullYear() <= _options.maximumYear ) {
                 build( nextYear );
                 triggerOptionsEventWithData( "onNextYear", nextYear );
             }
@@ -9550,7 +9550,6 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _options.showDayNumberOrdinals = getDefaultBoolean( _options.showDayNumberOrdinals, true );
         _options.dragAndDropForEventsEnabled = getDefaultBoolean( _options.dragAndDropForEventsEnabled, true );
         _options.maximumEventsPerDayDisplay = getDefaultNumber( _options.maximumEventsPerDayDisplay, 3 );
-        _options.extraSelectableYearsAhead = getDefaultNumber( _options.extraSelectableYearsAhead, 100 );
         _options.exportEventsEnabled = getDefaultBoolean( _options.exportEventsEnabled, true );
         _options.manualEditingEnabled = getDefaultBoolean( _options.manualEditingEnabled, true );
         _options.showTimesInMainCalendarEvents = getDefaultBoolean( _options.showTimesInMainCalendarEvents, false );
@@ -9594,6 +9593,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _options.weekendDays = isInvalidOptionArray( _options.weekendDays, 0 ) ? [ 0, 6 ] : _options.weekendDays;
         _options.addYearButtonsInDatePickerMode = getDefaultBoolean( _options.addYearButtonsInDatePickerMode, false );
         _options.workingDays = isInvalidOptionArray( _options.workingDays, 0 ) ? [] : _options.workingDays;
+        _options.minimumYear = getDefaultNumber( _options.minimumYear, 1900 );
+        _options.maximumYear = getDefaultNumber( _options.maximumYear, 2100 );
 
         if ( isInvalidOptionArray( _options.visibleDays ) ) {
             _options.visibleDays = [ 0, 1, 2, 3, 4, 5, 6 ];